Adding in old transforms; work on croppad lazy transforms
