+++ monai/apps/detection/utils/detector_utils.py
@@ -17,7 +17,7 @@ import torch
 import torch.nn.functional as F
 from torch import Tensor
 
-from monai.transforms.croppad.array import SpatialPad
+from monai.transforms.croppad.old_array import SpatialPad
 from monai.transforms.utils import compute_divisible_spatial_size, convert_pad_mode
 from monai.utils import PytorchPadMode, ensure_tuple_rep
