+++ tests/test_lazy_utils.py
@@ -3,7 +3,8 @@ import unittest
 import numpy as np
 
 from monai.transforms.utils import create_rotate, create_scale, create_flip, create_rotate_90, create_shear
-from monai.transforms.lazy.utils import matrix_to_eulers, check_matrix, check_axes, check_unit_translate
+from monai.transforms.lazy.utils import matrix_to_eulers, check_matrix, check_axes, check_unit_translate, \
+    get_scaling_factors
 
 import torch
 
@@ -31,33 +32,33 @@ class TestMatrixToEulers(unittest.TestCase):
 class TestCheckMatrix(unittest.TestCase):
 
     TEST_CASES = [
-        (np.eye(4), (True, True)),
-        (create_scale(3, (1.0, 1.0, 1.0)), (True, True)),
-        (create_scale(3, (1.1, 1.0, 1.0)), (True, False)),
-        (create_scale(3, (1.0, 1.1, 1.0)), (True, False)),
-        (create_scale(3, (1.0, 1.0, 1.1)), (True, False)),
-        (create_scale(3, (0.9, 1.0, 1.0)), (True, False)),
-        (create_scale(3, (1.0, 0.9, 1.0)), (True, False)),
-        (create_scale(3, (1.0, 1.0, 0.9)), (True, False)),
-        (create_scale(3, (0.9, 0.9, 0.9)), (True, False)),
-        (create_scale(3, (1.1, 1.1, 1.1)), (True, False)),
-        (create_flip(3, 0), (True, True)),
-        (create_flip(3, 1), (True, True)),
-        (create_flip(3, 2), (True, True)),
-        (create_flip(3, (0, 1)), (True, True)),
-        (create_flip(3, (0, 2)), (True, True)),
-        (create_flip(3, (1, 2)), (True, True)),
-        (create_flip(3, (0, 1, 2)), (True, True)),
-        (create_rotate_90(3, (1, 2), 0), (True, True)),
-        (create_rotate_90(3, (1, 2), 1), (True, True)),
-        (create_rotate_90(3, (1, 2), 2), (True, True)),
-        (create_rotate_90(3, (1, 2), 3), (True, True)),
-        (create_rotate(3, (0, 0, torch.pi / 2)), (True, True)),
-        (create_rotate(3, (0, 0, torch.pi)), (True, True)),
-        (create_rotate(3, (0, 0, 3 * torch.pi / 2)), (True, True)),
-        (create_rotate(3, (0, 0, 2 * torch.pi)), (True, True)),
-        (create_rotate(3, (0, 0, torch.pi / 4)), (False, True)),
-        (create_shear(3, 2, 0.5), (False, False))
+        (np.eye(4), (True, True, True)),
+        (create_scale(3, (1.0, 1.0, 1.0)), (True, True, True)),
+        (create_scale(3, (1.1, 1.0, 1.0)), (True, False, True)),
+        (create_scale(3, (1.0, 1.1, 1.0)), (True, False, True)),
+        (create_scale(3, (1.0, 1.0, 1.1)), (True, False, True)),
+        (create_scale(3, (0.9, 1.0, 1.0)), (True, False, True)),
+        (create_scale(3, (1.0, 0.9, 1.0)), (True, False, True)),
+        (create_scale(3, (1.0, 1.0, 0.9)), (True, False, True)),
+        (create_scale(3, (0.9, 0.9, 0.9)), (True, False, True)),
+        (create_scale(3, (1.1, 1.1, 1.1)), (True, False, True)),
+        (create_flip(3, 0), (True, True, True)),
+        (create_flip(3, 1), (True, True, True)),
+        (create_flip(3, 2), (True, True, True)),
+        (create_flip(3, (0, 1)), (True, True, True)),
+        (create_flip(3, (0, 2)), (True, True, True)),
+        (create_flip(3, (1, 2)), (True, True, True)),
+        (create_flip(3, (0, 1, 2)), (True, True, True)),
+        (create_rotate_90(3, (1, 2), 0), (True, True, True)),
+        (create_rotate_90(3, (1, 2), 1), (True, True, True)),
+        (create_rotate_90(3, (1, 2), 2), (True, True, True)),
+        (create_rotate_90(3, (1, 2), 3), (True, True, True)),
+        (create_rotate(3, (0, 0, torch.pi / 2)), (True, True, True)),
+        (create_rotate(3, (0, 0, torch.pi)), (True, True, True)),
+        (create_rotate(3, (0, 0, 3 * torch.pi / 2)), (True, True, True)),
+        (create_rotate(3, (0, 0, 2 * torch.pi)), (True, True, True)),
+        (create_rotate(3, (0, 0, torch.pi / 4)), (False, True, True)),
+        (create_shear(3, 2, 0.5), (False, False, False))
     ]
 
     def test_check_matrix_cases(self):
@@ -99,19 +100,70 @@ class TestCheckUnitTranslate(unittest.TestCase):
 class TestCheckAxes(unittest.TestCase):
 
     TEST_CASES = [
-        (np.eye(4), ((0, 1), (1, 1), (2, 1))),
-        (create_rotate(3, (0, 0, torch.pi / 2)), ((1, -1), (0, 1), (2, 1))),
-        (create_rotate(3, (0, torch.pi / 2, 0)), ((2, 1), (1, 1), (0, -1))),
-        (create_rotate(3, (torch.pi / 2, 0, 0)), ((0, 1), (2, -1), (1, 1))),
-        (create_rotate(3, (0, 0, torch.pi)), ((0, -1), (1, -1), (2, 1))),
-        (create_rotate(3, (0, torch.pi, 0)), ((0, -1), (1, 1), (2, -1))),
-        (create_rotate(3, (torch.pi, 0, 0)), ((0, 1), (1, -1), (2, -1))),
+        (np.eye(4), (tuple(), (0, 1, 2))),
+        (create_rotate(3, (0, 0, torch.pi / 2)), ((0,), (1, 0, 2))),
+        (create_rotate(3, (0, torch.pi / 2, 0)), ((2,), (2, 1, 0))),
+        (create_rotate(3, (torch.pi / 2, 0, 0)), ((1,), (0, 2, 1))),
+        (create_rotate(3, (0, 0, torch.pi)), ((0, 1), (0, 1, 2))),
+        (create_rotate(3, (0, torch.pi, 0)), ((0, 2), (0, 1, 2))),
+        (create_rotate(3, (torch.pi, 0, 0)), ((1, 2), (0, 1, 2))),
     ]
 
     def test_check_axes_cases(self):
         for i_c, c in enumerate(self.TEST_CASES):
-            with self.subTest(f"{i_c}"):
+            with self.subTest(i_c):
                 self._test_check_axes(*c)
 
     def _test_check_axes(self, matrix, expected):
-        self.assertEqual(check_axes(matrix), expected)#
+        self.assertEqual(check_axes(matrix), expected)
+
+
+class TestGetScalingFactors(unittest.TestCase):
+
+    TEST_CASES = [
+        (np.eye(4), (1, 1, 1)),
+        (create_scale(3, (1.5, 2.5, 3.5)), (1.5, 2.5, 3.5)),
+        (create_rotate_90(3, (0, 1), 1) @ create_scale(3, (1.5, 2.5, 3.5)), (2.5, -1.5, 3.5)),
+        (create_rotate_90(3, (0, 1), 2) @ create_scale(3, (1.5, 2.5, 3.5)), (-1.5, -2.5, 3.5)),
+        (create_rotate_90(3, (0, 1), 3) @ create_scale(3, (1.5, 2.5, 3.5)), (-2.5, 1.5, 3.5)),
+        (create_scale(3, (1.5, 2.5, 3.5)) @ create_rotate_90(3, (0, 1), 1), (1.5, -2.5, 3.5)),
+        (create_scale(3, (1.5, 2.5, 3.5)) @ create_rotate_90(3, (0, 1), 2), (-1.5, -2.5, 3.5)),
+        (create_scale(3, (1.5, 2.5, 3.5)) @ create_rotate_90(3, (0, 1), 3), (-1.5, 2.5, 3.5)),
+        (create_flip(3, tuple()) @ create_scale(3, (1.5, 2.5, 3.5)), (1.5, 2.5, 3.5)),
+        (create_flip(3, (0,)) @ create_scale(3, (1.5, 2.5, 3.5)), (-1.5, 2.5, 3.5)),
+        (create_flip(3, (1,)) @ create_scale(3, (1.5, 2.5, 3.5)), (1.5, -2.5, 3.5)),
+        (create_flip(3, (2,)) @ create_scale(3, (1.5, 2.5, 3.5)), (1.5, 2.5, -3.5)),
+        (create_flip(3, (0, 1)) @ create_scale(3, (1.5, 2.5, 3.5)), (-1.5, -2.5, 3.5)),
+        (create_flip(3, (0, 2)) @ create_scale(3, (1.5, 2.5, 3.5)), (-1.5, 2.5, -3.5)),
+        (create_flip(3, (1, 2)) @ create_scale(3, (1.5, 2.5, 3.5)), (1.5, -2.5, -3.5)),
+        (create_flip(3, (0, 1, 2)) @ create_scale(3, (1.5, 2.5, 3.5)), (-1.5, -2.5, -3.5)),
+        (create_scale(3, (1.5, 2.5, 3.5)) @ create_flip(3, tuple()), (1.5, 2.5, 3.5)),
+        (create_scale(3, (1.5, 2.5, 3.5)) @ create_flip(3, (0,)), (-1.5, 2.5, 3.5)),
+        (create_scale(3, (1.5, 2.5, 3.5)) @ create_flip(3, (1,)), (1.5, -2.5, 3.5)),
+        (create_scale(3, (1.5, 2.5, 3.5)) @ create_flip(3, (2,)), (1.5, 2.5, -3.5)),
+        (create_scale(3, (1.5, 2.5, 3.5)) @ create_flip(3, (0, 1)), (-1.5, -2.5, 3.5)),
+        (create_scale(3, (1.5, 2.5, 3.5)) @ create_flip(3, (0, 2)), (-1.5, 2.5, -3.5)),
+        (create_scale(3, (1.5, 2.5, 3.5)) @ create_flip(3, (1, 2)), (1.5, -2.5, -3.5)),
+        (create_scale(3, (1.5, 2.5, 3.5)) @ create_flip(3, (0, 1, 2)), (-1.5, -2.5, -3.5)),
+        (create_rotate(3, (0, 0, 0)) @ create_scale(3, (1.5, 2.5, 3.5)), (1.5, 2.5, 3.5)),
+        (create_rotate(3, (0, 0, torch.pi / 2)) @ create_scale(3, (1.5, 2.5, 3.5)), (-2.5, 1.5, 3.5)),
+        (create_rotate(3, (0, 0, torch.pi)) @ create_scale(3, (1.5, 2.5, 3.5)), (-1.5, -2.5, 3.5)),
+        (create_rotate(3, (0, 0, 3 * torch.pi / 2)) @ create_scale(3, (1.5, 2.5, 3.5)), (2.5, -1.5, 3.5)),
+    ]
+
+    def test_get_scaling_factors(self):
+        for i_c, c in enumerate(self.TEST_CASES):
+            with self.subTest(i_c):
+                self._test_get_scaling_factors(*c)
+
+    def _test_get_scaling_factors(self, matrix, expected):
+        actual = get_scaling_factors(matrix)
+        print(actual, expected)
+        self.assertEqual(actual, expected)
+
+    def test_matrix_stuff(self):
+        m1 = create_rotate_90(3, (0,1), 1)
+        m2 = create_scale(3, (1.5, 2.5, 3.5))
+        print(m1)
+        print(m2)
+        print(m2 @ m1)
