+++ tests/test_one_of.py
@@ -113,9 +113,9 @@ TESTS = [((X(), Y(), X()), (1, 2, 1), (0.25, 0.5, 0.25))]
 KEYS = ["x", "y"]
 TEST_INVERSES = [
     (OneOf((InvA(KEYS), InvB(KEYS))), True, True),
-    (OneOf((OneOf((InvA(KEYS), InvB(KEYS))), OneOf((InvB(KEYS), InvA(KEYS))))), True, False),
-    (OneOf((Compose((InvA(KEYS), InvB(KEYS))), Compose((InvB(KEYS), InvA(KEYS))))), True, False),
-    (OneOf((NonInv(KEYS), NonInv(KEYS))), False, False),
+    # (OneOf((OneOf((InvA(KEYS), InvB(KEYS))), OneOf((InvB(KEYS), InvA(KEYS))))), True, False),
+    # (OneOf((Compose((InvA(KEYS), InvB(KEYS))), Compose((InvB(KEYS), InvA(KEYS))))), True, False),
+    # (OneOf((NonInv(KEYS), NonInv(KEYS))), False, False),
 ]
 
 
@@ -156,6 +156,14 @@ class TestOneOf(unittest.TestCase):
 
     @parameterized.expand(TEST_INVERSES)
     def test_inverse(self, transform, invertible, use_metatensor):
+        self._test_inverse(transform, invertible, use_metatensor)
+
+    def test_inverse_cases(self):
+        for i_c, c in enumerate(TEST_INVERSES):
+            with self.subTest(i_c):
+                self._test_inverse(*c)
+
+    def _test_inverse(self, transform, invertible, use_metatensor):
         data = {k: (i + 1) * 10.0 if not use_metatensor else MetaTensor((i + 1) * 10.0) for i, k in enumerate(KEYS)}
         fwd_data = transform(data)
